backend/requirements.txt
pydantic-ai==0.2.15
pydantic==2.11.5
pandas==2.2.3
devtools==0.12.2
python-dotenv==1.1.0
fastapi==0.115.9
uvicorn==0.34.0
python-multipart==0.0.20
aiosqlite==0.21.0
ollama==0.5.2
motor==3.6.0
argon2-cffi==23.1.0
PyJWT==2.9.0